App structure. Receiving data from api.
