+++ src/components/app/app.css
@@ -1,5 +1,3 @@
-@import url("https://unpkg.com/leaflet@1.2.0/dist/leaflet.css");
-
 .App {
   text-align: center;
 }
