+++ src/components/map/map.css
@@ -1,0 +1,1 @@
+@import url("https://unpkg.com/leaflet@1.2.0/dist/leaflet.css");
